Update struct names to use exported types
The struct names in various files were updated to start with a capital letter which makes them exported and usable in other packages. This change allows these types to be used directly (declaration, parameters, return type).

Types that need initialization (and have a pseudo-constructor) and are exported through interfaces were not modified.